Move charts to Charts.js file
