+++ src/SleepRepository.js
@@ -100,43 +100,6 @@ class SleepRepository{
         return weekDate;
     }
 
-    displayWeeklySleepChart(chart, date, sHours, sqHours) {
-        if(chart != null){
-            chart.destroy()
-        }
-        var barColors = ['steelblue', 'steelblue', 'steelblue', 'steelblue', 'steelblue', 'steelblue', 'steelblue'];
-        var barColors2 = ['cadetblue', 'cadetblue', 'cadetblue', 'cadetblue', 'cadetblue', 'cadetblue', 'cadetblue'];
-        const data = {
-            labels: date,
-            datasets: [{
-                label: 'Sleep Hours',
-                backgroundColor: barColors,
-                data: sHours
-            },
-            {
-                label: 'Sleep Quality',
-                backgroundColor: barColors2,
-                data: sqHours
-            }]
-        }
-        const config = {
-            type: 'bar',
-            data: data,
-            options: {
-                sqHours,
-                sHours,
-                legend: {
-                    display: false
-                },
-                title: {
-                    display: true,
-                    text: 'Sleep Hours & Sleep Quality Per Day'
-                }
-            }
-        }
-
-        var chart = new Chart("sleepChart", config)
-    }
 };
